desktop: don't close edit dialog when save is cancelled
In the edit dialog, when OK is pressed, anotehr modal dialog opens and asks if
the template should be saved.  If the answer is negative, return to the edit
dialog rather than closing it abandoning the changes.

To abandon the changes, there is still the cancel buttton of the edit dialog.

Signed-off-by: Robert C. Helling <helling@atdotde.de>