desktop: don't update notes field if unchanged
User report: when switching focus between windows, the
cursor position gets lost. This is due to a note-edit
command being fired, which then overwrites the notes tab.

This should probably fixed in a better way (recognize that
the field is being edited and not emit a signal). However,
this is a quick fix for now.

Should this be extended to other fields?

Fixes #3365

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>