cleanup: avoid pos() member
I don't see the value of deprecating pos(), but whatever, this has
worked for a while now - at least in all Qt versions we still support.
So instead of having a simple three letter shorthand, we need to
explicitly convert the result of position().

Also, convert the one use of localPosition().

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>