profile: only keep pointer to plot_info in ruler objects
Now this one was strange:

The ruler items keep a copy of the plot_info struct. However,
only a shallow copy is made (the actual plit data is not copied).
This means that the data is only valid as long as the source
plot_info is valid. But if that is guaranteed, we simply can
keep a pointer instead of the full object.

I wonder if it wouldn't be better still to keep a pointer to
the profile and query that for the plot info?

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>