cleanup: cleanup profile tooltip code
This was a mess of plain pointers. For example to generate
the "entryToolTip", an item in the "toolTips" vector was
created, then deep-copied and then the "toolTips" vector
was cleared.

Instead, replace the "addToolTip" by a "makeToolTip" function
and replace plain pointers by unique_ptrs. Makes this thing
a bit more manageable.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>