profile: directly format QString for tooltip
Instead of rendering in a membuffer, directly format into the
QString. We have the qasprintf_loc() function, which is equivalent
to put_format_loc(). Thus, this is a manual search/replace job.

However, this means that the formatting is done in UTF-8 and
then piecewise transformed into QString's (inane) UTF-16.
In the future we might think about directly using QString's
formatting function.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>